Add files via upload
Analyzing similarity on French Articles from a given Database in order to group them by theme. 
The Database is read and put in a list of a dictionaries... 
Every Database has its own tables so in order to make this code work, one needs to adjust the data according to their own Database.